+++ tests/smoke-profiles.sh
@@ -6,6 +6,7 @@ source /opt/gcc12-toolset/enable full
 [[ $CXX == /opt/gcc12-toolset/root/usr/bin/g++ ]]
 [[ :$LD_LIBRARY_PATH: == *:/opt/gcc12-toolset/root/usr/lib64/binutils:* ]]
 [[ :$LD_LIBRARY_PATH: == *:/opt/gcc12-toolset/root/usr/lib64:* ]]
+[[ :$LD_LIBRARY_PATH: == *:/opt/gcc12-toolset/root/usr/lib:* ]]
 "$CXX" --version | grep -F '12.2.1' >/dev/null
 ld --version | grep -F '2.36.1' >/dev/null
 
@@ -15,6 +16,7 @@ source /opt/gcc12-toolset/enable compat
 [[ $(command -v g++) == /opt/gcc12-toolset/profiles/compat/bin/g++ ]]
 [[ :$LD_LIBRARY_PATH: == *:/opt/gcc12-toolset/root/usr/lib64/binutils:* ]]
 [[ :$LD_LIBRARY_PATH: != *:/opt/gcc12-toolset/root/usr/lib64:* ]]
+[[ :$LD_LIBRARY_PATH: != *:/opt/gcc12-toolset/root/usr/lib:* ]]
 ld --version | grep -F '2.36.1' >/dev/null
 
 work=$(mktemp -d)